+++ gemrb/core/GUI/TextArea.h
@@ -79,7 +79,7 @@ public:
 	/** Sets up auto scrolling (chapter text) */
 	void SetupScroll();
 	/** Per Pixel scrolling */
-	void ScrollToY(unsigned long y, Control* sender);
+	void ScrollToY(int y, Control* sender);
 
 	/** Returns total height of the text */
 	int GetRowHeight() const;
